feat: improve route delete user

+++ src/controllers/users/routes/delete.ts
@@ -5,6 +5,18 @@ import { Request } from 'express';
 async function del(req: Request, res: ResponseUserId) {
 	const id = res.locals.userId;
 
+	const user = await prisma.user.findUnique({
+		where: {
+			id,
+		},
+	});
+
+	if (!user) {
+		return res
+			.status(400)
+			.send({ error: true, message: 'This user does not exist' });
+	}
+
 	try {
 		await prisma.user.delete({
 			where: {
@@ -12,12 +24,13 @@ async function del(req: Request, res: ResponseUserId) {
 			}
 		});
 
-		return res.status(200);
-	} catch(err) {
-		console.log(err);
 		return res
-			.status(400)
-			.send({ error: true, message: 'Error deleting user' });
+			.status(200)
+			.send({ success: true, message: 'User deleted successfully' });
+	} catch(err) {
+		return res
+			.status(500)
+			.send({ error: true, message: 'Internal server error' });
 	}
 }
